+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Dont instantiate socket connection unless we need to
+- Disable loguru logging by default, require user to enable and configure it (#4)
+- Added documentation on how to configure loguru (#4)
 
 ## [1.1.3] - 2023-08-24